% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{slope}
\alias{slope}
\title{Slope of a time series interval}
\usage{
slope(y)
}
\arguments{
\item{y}{time series interval values (numeric vector)}
}
\value{
slope (numeric scalar)
}
\description{
Calculates the ordinary least squares slope of the time series regressed
against the sequence 1:M where M is the length of the interval.
}
\examples{
slope(1:10) # 1
slope(runif(10000)) # ~0
}
