% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tsfeats}
\alias{tsfeats}
\title{Extract time series features from intervals}
\usage{
tsfeats(tslen, tsmat, tsints)
}
\arguments{
\item{tslen}{length of each time series}

\item{tsmat}{a matrix of time series data}

\item{tsints}{interval start and end indices}
}
\description{
Extract time series features from intervals
}
