% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{extract_features}
\alias{extract_features}
\title{Extract features from time series}
\usage{
extract_features(
  x,
  tsid,
  intervals,
  funs = list(mean = mean, sd = sd, slope = slope)
)
}
\arguments{
\item{x}{time series (data.frame)}

\item{tsid}{name of time series identifier column in \code{x} (character scalar)}

\item{intervals}{start and end indices of intervals (2 column integer matrix;
see \code{\link{sample_intervals}})}

\item{funs}{list of summary functions (\code{\link{mean}}, \code{\link{sd}},
and \code{\link{slope}} by default)}
}
\value{
an MxN matrix where M is the length of \code{intervals} and N is the
length of \code{funs}.
}
\description{
Calculate summary statistics of randomly sampled intervals in a time series
}
\examples{
ts_len <- 100
ts_dat <- data.frame(id = rep(1:3, each = ts_len),
                     val = 1:ts_len)
ints <- sample_intervals(ts_len)
extract_features(ts_dat, "id", ints)
}
